requirements.txt
torch # NOTE: you need cuda enabled torch
safetensors # for saving & loading model weights
datasets # for handling data
tokenizers 
huggingface_hub # for storing model weights in cloud
lightning # for organizing training
pillow # for image pre-processing
matplotlib # for plotting
